# openapi/agent-runtime-api.yaml
openapi: 3.1.0
info:
  title: Agent Runtime API
  version: 0.1.0
  description: Minimal HTTP surface for the schema-first Rust agent runtime.
...
paths:
...
  /chat/turn:
    post:
      operationId: streamChatTurn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/chat-turn-request.schema.json"
      responses:
        "200":
          description: Server-sent event stream of chat turn events.
          content:
            text/event-stream:
              schema:
                type: string
                description: Each SSE data frame is a ChatTurnEvent JSON object from chat-turn-event.schema.json.
        "400":
          description: Request body failed JSON or schema validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Chat turn failed before streaming started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/resume:
    post:
      operationId: streamChatResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/chat-resume-request.schema.json"
      responses:
        "200":
          description: Server-sent event stream of resumed chat turn events.
          content:
            text/event-stream:
              schema:
                type: string
                description: Each SSE data frame is a ChatTurnEvent JSON object from chat-turn-event.schema.json.
        "400":
          description: Request body failed JSON or schema validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Chat resume failed before streaming started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workflows/run:
    post:
      operationId: runWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/workflow-run-request.schema.json"
      responses:
        "200":
          description: Completed workflow DAG invocation result.
          content:
            application/json:
              schema:
                $ref: "../schemas/workflow-run-result.schema.json"
        "400":
          description: Request body failed JSON or schema validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Workflow invocation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /runs:
    get:
      operationId: listRuns
      parameters:
        - name: agent_id
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Persisted run records, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRunRecord"
        "500":
          description: Run listing failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}:
    get:
      operationId: inspectRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^run_"
      responses:
        "200":
          description: Persisted run record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunRecord"
        "404":
          description: Run was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}/trace:
    get:
      operationId: inspectRunTrace
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^run_"
      responses:
        "200":
          description: Persisted run trace.
          content:
            application/json:
              schema:
                $ref: "../schemas/trace.schema.json"
        "404":
          description: Trace was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}/events:
    get:
      operationId: streamRunTraceEvents
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^run_"
      responses:
        "200":
          description: Server-sent event stream of active run events when the run is live, otherwise persisted trace events.
          content:
            text/event-stream:
              schema:
                type: string
                description: Each SSE data frame is a TraceEvent JSON object from trace.schema.json.
        "404":
          description: Trace was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^run_"
      responses:
        "200":
          description: Cancellation request result. Active runs receive cancellation and running records persist a cancellation intent under metadata.control.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelRunResponse"
        "404":
          description: Run was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}/replay:
    post:
      operationId: replayRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^run_"
      responses:
        "200":
          description: Live replay result generated from the persisted trace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayExecutionResponse"
        "500":
          description: Replay failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tools/{tool_name}/call:
    post:
      operationId: callTool
      parameters:
        - name: tool_name
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  default: {}
              additionalProperties: false
      responses:
        "200":
          description: Tool call output envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallResponse"
        "500":
          description: Tool call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proposals:
    get:
      operationId: listProposals
      parameters:
        - name: run_id
          in: query
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Persisted proposal envelopes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "../schemas/proposal-envelope.schema.json"
    post:
      operationId: createProposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalCreateRequest"
      responses:
        "200":
          description: Created proposal envelope.
          content:
            application/json:
              schema:
                $ref: "../schemas/proposal-envelope.schema.json"
        "403":
          description: Proposal creation was denied by policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Proposal creation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proposals/{proposal_id}:
    get:
      operationId: inspectProposal
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Proposal envelope.
          content:
            application/json:
              schema:
                $ref: "../schemas/proposal-envelope.schema.json"
        "404":
          description: Proposal was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proposals/{proposal_id}/decision:
    post:
      operationId: decideProposal
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalDecisionRequest"
      responses:
        "200":
          description: Approval decision and updated proposal envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalDecisionResponse"
        "500":
          description: Proposal decision failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proposals/{proposal_id}/apply:
    post:
      operationId: applyProposal
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Proposal apply result and updated proposal envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalActionResponse"
        "403":
          description: Proposal apply was denied by policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Proposal apply failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CancelRunResponse:
      type: object
      required: [cancellation_requested, message, run_id]
      properties:
        cancellation_requested:
          type: boolean
          description: True when an active run was signaled or a persisted running record was marked for cancellation.
        message:
          type: string
        run_id:
          type: string
          pattern: "^run_"
        status:
          type: string
          enum:
            - running
            - completed
            - skipped
            - failed
            - cancelled
            - timed_out
            - abandoned
      additionalProperties: false
...
    ToolCallResponse:
      type: object
      required: [tool, output]
      properties:
        tool:
          type: string
          minLength: 1
        output:
          type: object
      additionalProperties: false
...
    ProposalDecisionResponse:
      type: object
      required: [decision, proposal]
      properties:
        decision:
          $ref: "../schemas/approval-decision.schema.json"
        proposal:
          $ref: "../schemas/proposal-envelope.schema.json"
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        details:
          type: object
          description: Optional structured error details. Policy denials include the hook event and proposal context.
      additionalProperties: false